+++ ruststep/src/ast/de/mod.rs
@@ -1,0 +1,9 @@
+//! Implementation of [serde::de] traits for AST structs
+
+mod parameter;
+mod record;
+mod value;
+
+pub use parameter::*;
+pub use record::*;
+pub use value::*;
